revise container and component naming conventions
