--- client/components/UserAnswerInputBTN.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const UserAnswerInputBtn = () => (
-  <button>User Input</button>
-);
-
-export default UserAnswerInputBtn;
