--- client/components/UserAnswerInputFalseBTN.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const UserAnswerInputFalseBtn = () => (
-  <button> B: False </button>
-);
-
-export default UserAnswerInputFalseBtn;
