--- client/components/UserAnswerInputTrueBTN.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const UserAnswerInputTrueBtn = () => (
-  <button> A: True </button>
-);
-
-export default UserAnswerInputTrueBtn;
