+++ client/components/WorkingMemoryQuestionDisplay.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Button from '@material-ui/core/Button';
 
-const WMQuestionDisplay = props => {
+const WorkingMemoryQuestionDisplay = props => {
 
   return (
     <div>
@@ -80,4 +80,4 @@ const WMQuestionDisplay = props => {
   )
 };
 
-export default WMQuestionDisplay;
+export default WorkingMemoryQuestionDisplay;
