+++ client/containers/App.jsx
@@ -4,8 +4,8 @@ import { connect } from 'react-redux';
 
 import Header from './../components/Header.jsx';
 import Authentication from './../components/Authentication.jsx';
-import MainDisplay from "./MainDisplay.jsx"
-import UserDemographics from "./UserDemographics.jsx"
+import Main from "./Main.jsx"
+import UserDemographics from "./Demographics.jsx"
 import Login from './Login.jsx';
 import {BrowserRouter as Router, Switch, Route} from 'react-router-dom'
 
@@ -27,7 +27,7 @@ class App extends Component {
         <Switch>
           <Route path='/login' component={Login} />
             {/* <Authentication> */}
-          <Route path='/main' component={MainDisplay}/>
+          <Route path='/main' component={Main}/>
             {/* </Authentication> */}
         </Switch>
      </div>
